+++ walle/template/eslint.config.js
@@ -1,3 +1,0 @@
-import { eslintWalleConfigs } from "./src/@walle/eslint.preset.js";
-
-export default [...eslintWalleConfigs];